More validation
I need to work on the phrasing of my error messages...

+++ TPetshop2019.Core/ApplicationServices/Services/OwnerService.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using System.IO;
 using System.Linq;
 using System.Runtime;
 using TPetshop2019.Core.DomainServices;
@@ -11,16 +12,21 @@ namespace TPetshop2019.Core.ApplicationServices.Services
     public class OwnerService: IOwnerService
     {
         private readonly IOwnerRepository _ownerRepo;
-        private readonly IValidateIdService _valiService;
+        private readonly IValidateIdService _validateIdService;
 
-        public OwnerService(IOwnerRepository ownerRepo, IValidateIdService valiService)
+        public OwnerService(IOwnerRepository ownerRepo, IValidateIdService validateIdService)
         {
             this._ownerRepo = ownerRepo;
-            this._valiService = valiService;
+            this._validateIdService = validateIdService;
         }
 
         public Owner NewOwner(string firstName, string lastName, string address, string phoneNr, string email)
         {
+            if (string.IsNullOrEmpty(firstName) || string.IsNullOrEmpty(lastName))
+            {
+                throw new InvalidDataException("The owner needs both a first and last name");
+            }
+
             Owner owner = new Owner
             {
                 Address = address,
@@ -34,26 +40,36 @@ namespace TPetshop2019.Core.ApplicationServices.Services
 
         public Owner CreateOwner(Owner owner)
         {
+            if (owner == null)
+            {
+                throw new InvalidDataException("The owner object is null and therefor invalid");
+            }
             return _ownerRepo.CreateOwner(owner);
         }
 
         public Owner ReadOwner(int id)
         {
-            return _ownerRepo.GetOwners().ToList().FirstOrDefault(owner => owner.Id == id);
+            _validateIdService.ValidateId(id);
+
+            var ownerToGet = _ownerRepo.GetOwners().ToList().FirstOrDefault(owner => owner.Id == id);
+
+            if (ownerToGet == null)
+            {
+                throw new InvalidDataException($"No owner with the id: {id} found");
+            }
+
+            return ownerToGet;
         }
 
         public Owner UpdateOwner(int id, string firstName, string lastName, string address, string phoneNr, string email)
         {
             Owner ownerToUpdate = ReadOwner(id);
-
-            if (ownerToUpdate != null)
-            {
-                ownerToUpdate.FirstName = firstName;
-                ownerToUpdate.LastName = lastName;
-                ownerToUpdate.Email = email;
-                ownerToUpdate.Address = address;
-                ownerToUpdate.PhoneNumber = phoneNr;
-            }
+           
+            ownerToUpdate.FirstName = firstName;
+            ownerToUpdate.LastName = lastName;
+            ownerToUpdate.Email = email;
+            ownerToUpdate.Address = address;
+            ownerToUpdate.PhoneNumber = phoneNr;
             return ownerToUpdate;
         }
         public Owner MakeUpdatedOwner(Owner ownerToUpdate)
@@ -68,6 +84,10 @@ namespace TPetshop2019.Core.ApplicationServices.Services
         }
         public Owner DeleteOwner(Owner owner)
         {
+            if (owner == null)
+            {
+                throw new InvalidDataException($"Something went wrong and the owner was null");
+            }
             return _ownerRepo.DeleteOwner(owner);
         }
